algo3_finales-master/final-04-02-2021/final-4-2-21/src/modelo/Usuario.java
package modelo;

import modelo.Alertas.Alerta;

public class Usuario {

    private int timbrazosRecibidos = 0;
    private int alertasDeMovimientoRecibidas = 0;
    private final String nombre;

    public Usuario(String nombre) {
        this.nombre = nombre;
    }

    public void notificarMovimiento(Alerta alerta) {
        alertasDeMovimientoRecibidas+=1;
    }

    public void notificarTimbre(){
        timbrazosRecibidos+=1;
    }

    public int getMovimientosNotificados(){
        return alertasDeMovimientoRecibidas;
    }

    public int getTimbrazosNotificados(){
        return timbrazosRecibidos;
    }
}
